// packages/components/src/types.ts
export type commonTypes = {
    variant?: 'flat' | 'solid' | 'outlined' | 'soft'
    color?:
      | 'primary'
      | 'secondary'
      | 'tertiary'
      | 'info'
      | 'success'
      | 'warning'
      | 'danger'
      | 'default'
      | 'light'
      | 'dark'
    size?: 'sm' | 'base' | 'lg' | 'xl'
    radius?: 'none' | 'soft' | 'round'
  }
  